+++ mne/inverse_sparse/mxne_inverse.py
@@ -422,6 +422,8 @@ def tf_mixed_norm(evoked, forward, noise_cov, alpha_space, alpha_time,
         is less than weights_min.
     pca: bool
         If True the rank of the data is reduced to true dimension.
+    debias: bool
+        Remove coefficient amplitude bias due to L1 penalty.
     wsize: int
         Length of the STFT window in samples (must be a multiple of 4).
     tstep: int
@@ -431,12 +433,10 @@ def tf_mixed_norm(evoked, forward, noise_cov, alpha_space, alpha_time,
         Length of time window used to take care of edge artifacts in seconds.
         It can be one float or float if the values are different for left
         and right window length.
-    debias: bool
-        Remove coefficient amplitude bias due to L1 penalty.
     return_residual : bool
         If True, the residual is returned as an Evoked instance.
-    verbose: bool
-        Verbose output or not.
+    verbose : bool, str, int, or None
+        If not None, override default verbose level (see mne.verbose).
 
     Returns
     -------