+++ mne/realtime/mockclient.py
@@ -12,12 +12,10 @@ from ..event import find_events
 class MockRtClient(object):
     """Mock Realtime Client
 
-    Attributes
+    Parameters
     ----------
     raw : instance of Raw object
         The raw object which simulates the RtClient
-    info : dict
-        Measurement info.
     verbose : bool, str, int, or None
         If not None, override default verbose level (see mne.verbose).
     """
@@ -161,15 +159,33 @@ class MockRtClient(object):
             return None
 
     def register_receive_callback(self, x):
-        """API boilerplate"""
+        """API boilerplate
+
+        Parameters
+        ----------
+        x : None
+            Not used.
+        """
         pass
 
     def start_receive_thread(self, x):
-        """API boilerplate"""
+        """API boilerplate
+
+        Parameters
+        ----------
+        x : None
+            Not used.
+        """
         pass
 
     def unregister_receive_callback(self, x):
-        """API boilerplate"""
+        """API boilerplate
+
+        Parameters
+        ----------
+        x : None
+            Not used.
+        """
         pass
 
     def _stop_receive_thread(self):