+++ mne/time_frequency/__init__.py
@@ -5,7 +5,7 @@ from .tfr import (single_trial_power, morlet, tfr_morlet, cwt_morlet,
                   AverageTFR, tfr_multitaper, read_tfrs, write_tfrs)
 from .psd import compute_raw_psd, compute_epochs_psd
 from .csd import CrossSpectralDensity, compute_epochs_csd
-from .ar import yule_walker, ar_raw, iir_filter_raw
+from .ar import iir_filter_raw, fit_iir_model_raw
 from .multitaper import dpss_windows, multitaper_psd
 from .stft import stft, istft, stftfreq
 from ._stockwell import tfr_stockwell